Added UEs connected and Ave Throughput Per User (displayed in the KPIs widget)
Added Time Series Charts for Number of UEs Connected and Ave Throughput Per User.
Color code for those charts: Green if throughput is >= 2, yellow if it is >=1 and < 2, red if it is < 1

Added EDX Apis to get 3 rsrps values for celltower locations. Displayed as rsrp1, rsrp2 and rsrp3 in the KPI Widge